Split ban modals, add backup CLI, fix demo IDs
- Split combined ban modal into separate create and edit components
- Add CLI backup command to sync server logs via rsync
- Change demo_id protobuf field from int64 to int32
- Fix report search filter column and wiki route loading
- Remove debug transport interceptor and update CSP
- Regenerate frontend and backend protobuf code

+++ internal/ban/report.go
@@ -47,7 +47,7 @@ type RequestReportCreate struct {
 	Description     string
 	Reason          reason.Reason
 	ReasonText      string
-	DemoID          int64
+	DemoID          int32
 	DemoTick        int32
 	PersonMessageID int64
 }
@@ -85,7 +85,7 @@ type Report struct {
 	ReasonText      string
 	Deleted         bool
 	DemoTick        int32
-	DemoID          int64
+	DemoID          int32
 	PersonMessageID int64
 	CreatedOn       time.Time
 	UpdatedOn       time.Time
